stuff is happening

+++ hawk/src/util.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Eq, PartialEq)]
 pub struct Pos {
-  pub row: u8,
-  pub column: u8,
+  pub row: u16,
+  pub column: u16,
 }
 
 impl Pos {
-  pub fn new(row: u8, column: u8) -> Self {
+  pub fn new(row: u16, column: u16) -> Self {
     Pos { row, column }
   }
 }
